--- a/griffis_soccer_analysis/similarity_scores.py
+++ b/griffis_soccer_analysis/similarity_scores.py
@@ -3,7 +3,6 @@ import matplotlib.pyplot as plt
 import seaborn as sns
 import pandas as pd
 import numpy as np
-import statistics
 from sklearn.metrics.pairwise import cosine_similarity
 import warnings
 warnings.filterwarnings("ignore")