By analogy with https://github.com/griffisben/griffis_soccer_analysis/commit/67cd8ba2341002aa2934a3c6b22f0fee8f775820

i replace pd.read_parquet on read_parquet and changed code inside for:
team_similarity function (Speed up 6s -> 0.5s)
leaguw_similarity (Speed up 330 ms -> 2 ms)

![Снимок экрана от 2023-06-21 13-48-26](https://github.com/griffisben/griffis_soccer_analysis/assets/49508936/5b12d382-c721-4cf0-9643-b6e13ce52a29)